requirements_lock.txt
absl-py==1.3.0
cadquery-ocp==7.5.3.0
cadquery==2.2.0b2
casadi==3.5.6rc2
ezdxf==1.0.0rc2
multimethod==1.9
nlopt==2.7.1
nptyping==2.0.1
numpy==1.23.5
path==16.5.0
scipy==1.9.3
typish==1.9.3
